improvement: finish our User and Token model :heart_eyes:
In the User model I just create a class with five methods for help us when we are create a user
model.

addHook method to create a hash of the password before save the data on
db.

checkPassword method to check if the password insert by the user is the same saved in the
database.

generateJWT method for creating the auth token using jwt pkg.

genaratePasswordReset
method for generate a password reset token using crypto pkg and calculates an expiry time (set for
1h).

getVerificationToken method used for generate a token and creates, returns an instance of
the Token model.

closes #10