improvement(passwordcontroller.js): implementing reset and resetPassword
In the password controller I work in 4 methods.

**recover**
Call the **generatePasswordReset**
method if the user's email is found, after that a reset link is created.

**reset**That method
check the user object from the database using the password reset token and verifying if is valid
compare eith the `Date.now()`

**resetPassword**
This method verify if the token still valid, by
comparing the date that was created with `Date.now()`.

If the token still valid the user password
will be update.